feat: tsconfig.json에 declaration, outDir 설정 추가
- tsc 컴파일 결과가 ./dist 파일에 저장되도록 수정
- 타입 선언 파일도 생성되도록 수정
